split zabbix-get tests
