+++ .travis.yml
@@ -22,4 +22,6 @@ script:
   - docker logs docker_zabbix-agent-modbus_1
   - docker exec docker_zabbix-agent-modbus_1 sh -c "zabbix_get -s localhost -k agent.ping"
   - docker exec docker_zabbix-agent-modbus_1 sh -c "zabbix_get -s localhost -k modbus_read[172.16.238.2:5020,1,2,3,i,1,1]"
-  - python tests/modbus_docker_tests.py
+  - python tests/test_modbus.py
+  - python tests/test_modbus_errors.py
+  - python tests/test_modbus_64bit.py
